Adesão de outro modelo de organização de repositórios
