--- projeto/src/entities/grupamentoPessoa/ClienteRepository.java
@@ -1,9 +1,0 @@
-package entities.grupamentoPessoa;
-
-import java.io.IOException;
-
-import entities.grupamentoCliente.Cliente;
-
-public interface ClienteRepository {
-    public Cliente getClienteCpf(String cpf) throws IOException;
-}
