+++ projeto/src/negocio/ClienteRepository.java
@@ -1,4 +1,4 @@
-package entities.grupamentoCliente;
+package negocio;
 
 import java.io.IOException;
