+++ projeto/src/persistencia/SistemaDeArquivosColaboradorVendedorRepository.java
@@ -1,22 +1,23 @@
-package entities.sistemaDeArquivos;
+package persistencia;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 
-import entities.grupamentoCliente.Cliente;
-import entities.grupamentoContato.Contato;
-import entities.grupamentoEndereco.Endereco;
-import entities.grupamentoPessoa.ClienteRepository;
-import entities.grupamentoPessoa.EstadoCivil;
+import negocio.ColaboradorVendedor;
+import negocio.ColaboradorVendedorRepository;
+import negocio.Contato;
+import negocio.Endereco;
+import negocio.EstadoCivil;
 
-public class SistemaDeArquivosClienteRepository implements ClienteRepository {
+public class SistemaDeArquivosColaboradorVendedorRepository implements ColaboradorVendedorRepository {
 
     @Override
-    public Cliente getClienteCpf(String cpfParaBuscar) throws IOException {
+    public ColaboradorVendedor getColaboradorVendedorId(int idParaBuscar) throws IOException {
+
         FileReader r = null;
         BufferedReader bf = null;
-        String arquivo = "C:\\Users\\danie\\Documents\\gestaoConcessionaria\\projeto\\arquivos\\clientes.txt";
+        String arquivo = "C:\\Users\\danie\\Documents\\gestaoConcessionaria\\projeto\\arquivos\\colaboradoresVendedores.txt";
 
         try {
             r = new FileReader(arquivo);
@@ -26,9 +27,8 @@ public class SistemaDeArquivosClienteRepository implements ClienteRepository {
             String[] campos = new String[13];
 
             while ((linha = bf.readLine()) != null) {
-
                 campos = linha.split(";");
-                if (campos[11].equals(cpfParaBuscar)) {
+                if (Integer.parseInt(campos[14]) == idParaBuscar) {
                     String nome = campos[0];
                     String email = campos[1];
                     String telefone = campos[2];
@@ -42,12 +42,15 @@ public class SistemaDeArquivosClienteRepository implements ClienteRepository {
                     EstadoCivil estadoCivil = EstadoCivil.valueOf(campos[10].toUpperCase());
                     String cpf = campos[11];
                     String rg = campos[12];
+                    double salarioBase = Double.parseDouble(campos[13]);
+                    int id = Integer.parseInt(campos[14]);
 
                     Contato contato = new Contato(telefone, email);
                     Endereco endereco = new Endereco(cidade, estado, cep, rua, numero, bairro, complemento);
 
-                    return new Cliente(nome, contato, endereco, estadoCivil, cpf, rg);
+                    return new ColaboradorVendedor(nome, contato, endereco, estadoCivil, cpf, rg, salarioBase, id);
                 }
+
             }
         } catch (IOException e) {
             System.out.println(e.getMessage());
